src/requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Deprecated==1.2.13
Flask==2.0.2
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-appengine-logging==0.1.5
google-cloud-audit-log==0.1.1
google-cloud-bigquery==2.28.0
google-cloud-bigquery-storage==2.9.0
google-cloud-core==2.0.0
google-cloud-logging==2.6.0
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.2
opentelemetry-api==1.5.0
packaging==21.0
pandas==1.3.3
proto-plus==1.19.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
